add database crmdb and setup

+++ src/main/java/com/example/demo/dto/ClientResponse.java
@@ -1,17 +1,17 @@
 package com.example.demo.dto;
 
 public class ClientResponse {
-    private int id;
+    private Long id;
     private String name;
     private String email;
 
-    public ClientResponse(int id,String name,String email){
+    public ClientResponse(Long id,String name,String email){
         this.id = id;
         this.name = name;
         this.email = email;
     }
 
-    public int getId(){ return id; }
+    public Long getId(){ return id; }
     public String getName(){ return name; }
     public String getEmail(){ return email; }
 }
